Add compatibility layer for Windows (MSVC)
On Windows, std::filesystem::path::value_type is wchar_t.
Therefore we can't use path.c_str() as char* directly.

This changeset adds conversion function (wstring -> MB string) to mitigate this issue.

Note that this is not a ideal solution.  We should introduce path agnostic API for initialization of espeak-ng.