Update TRAINING writeup with up-to-date docker instructions
The suggested pytorch container release 22.03 has cuda libraries that are too old for the newer nvidia cards. 22.10 is the first version with support for the Lovelace architecture, so that is what I used and confirmed working. I also found it necessary to install v1.8.6 specifically of pytorch-lightning - newer versions would not work. Also added some tips on how to launch the container and how to deal with pathnames changing inside the container.